update store id from int to long

+++ TTO_MAIN/src/com/group/tto/main/dao/StoreDao.java
@@ -6,7 +6,7 @@ import com.group.tto.cmn.model.Store;
 import com.group.tto.main.common.StoreSearchCriteria;
 
 public interface StoreDao {
-	public Store getStoreById(int sid);
+	public Store getStoreById(Long sid);
